+++ src/main/java/prismatic/shards/stellarity/StellarityDatagen.java
@@ -6,6 +6,7 @@ import net.minecraft.core.RegistrySetBuilder;
 import net.minecraft.core.registries.Registries;
 import org.jetbrains.annotations.Nullable;
 import prismatic.shards.stellarity.datagen.*;
+import prismatic.shards.stellarity.datagen.dynamic.*;
 import prismatic.shards.stellarity.datagen.loot_table.BlockLootTableProvider;
 import prismatic.shards.stellarity.datagen.loot_table.ChestLootTableProvider;
 import prismatic.shards.stellarity.datagen.loot_table.FishingLootTableProvider;
@@ -14,15 +15,10 @@ import prismatic.shards.stellarity.datagen.tags.*;
 public class StellarityDatagen implements DataGeneratorEntrypoint {
 	@Override
 	public void buildRegistry(RegistrySetBuilder registryBuilder) {
-		Stellarity.LOGGER.info("Bootstrapping biomes");
 		registryBuilder.add(Registries.BIOME, BiomeProvider::bootstrap);
-		Stellarity.LOGGER.info("Bootstrapping configured features");
 		registryBuilder.add(Registries.CONFIGURED_FEATURE, ConfiguredFeatureProvider::bootstrap);
-		Stellarity.LOGGER.info("Bootstrapping placed features");
 		registryBuilder.add(Registries.PLACED_FEATURE, PlacedFeatureProvider::bootstrap);
-		Stellarity.LOGGER.info("Bootstrapping villager trades");
 		registryBuilder.add(Registries.VILLAGER_TRADE, VillagerTradeProvider::bootstrap);
-		Stellarity.LOGGER.info("Bootstrapping trade sets");
 		registryBuilder.add(Registries.TRADE_SET, TradeSetProvider::bootstrap);
 	}