+++ src/main/java/prismatic/shards/stellarity/datagen/DynamicRegistriesProvider.java
@@ -4,7 +4,9 @@ import net.fabricmc.fabric.api.datagen.v1.FabricPackOutput;
 import net.fabricmc.fabric.api.datagen.v1.provider.FabricDynamicRegistryProvider;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.core.registries.Registries;
+import net.minecraft.resources.RegistryDataLoader;
 import org.jspecify.annotations.NonNull;
+import prismatic.shards.stellarity.datagen.dynamic.BiomeProvider;
 
 import java.util.concurrent.CompletableFuture;