+++ src/main/java/prismatic/shards/stellarity/datagen/dynamic/PlacedFeatureProvider.java
@@ -1,4 +1,4 @@
-package prismatic.shards.stellarity.datagen;
+package prismatic.shards.stellarity.datagen.dynamic;
 
 import net.minecraft.core.HolderGetter;
 import net.minecraft.core.registries.Registries;
@@ -11,16 +11,15 @@ import prismatic.shards.stellarity.Stellarity;
 
 import java.util.List;
 
-public class PlacedFeatureProvider {
+public interface PlacedFeatureProvider {
+	ResourceKey<PlacedFeature> MAIN_ISLAND_RING = id("main_island/ring");
+	ResourceKey<PlacedFeature> MAIN_ISLAND_PORTAL_PLATFORM = id("main_island/portal_platform");
 
-	public static final ResourceKey<PlacedFeature> MAIN_ISLAND_RING = id("main_island/ring");
-	public static final ResourceKey<PlacedFeature> MAIN_ISLAND_PORTAL_PLATFORM = id("main_island/portal_platform");
-
-	public static ResourceKey<PlacedFeature> id(String name) {
+	static ResourceKey<PlacedFeature> id(String name) {
 		return Stellarity.key(Registries.PLACED_FEATURE, name);
 	}
 
-	public static void bootstrap(BootstrapContext<PlacedFeature> context) {
+	static void bootstrap(BootstrapContext<PlacedFeature> context) {
 
 		HolderGetter<ConfiguredFeature<?, ?>> configured = context.lookup(Registries.CONFIGURED_FEATURE);