+++ src/main/java/prismatic/shards/stellarity/tags/StellarityVillagerTradeTags.java
@@ -9,9 +9,83 @@ public interface StellarityVillagerTradeTags {
 	TagKey<VillagerTrade> ARMORER_LEVEL_1 = id("armorer/level_1");
 	TagKey<VillagerTrade> ARMORER_LEVEL_1_2 = id("armorer/level_1_2");
 	TagKey<VillagerTrade> ARMORER_LEVEL_2 = id("armorer/level_2");
+	TagKey<VillagerTrade> ARMORER_LEVEL_2_2 = id("armorer/level_2_2");
 	TagKey<VillagerTrade> ARMORER_LEVEL_3 = id("armorer/level_3");
 	TagKey<VillagerTrade> ARMORER_LEVEL_4 = id("armorer/level_4");
 	TagKey<VillagerTrade> ARMORER_LEVEL_5 = id("armorer/level_5");
+
+	TagKey<VillagerTrade> BUTCHER_LEVEL_1 = id("butcher/level_1");
+	TagKey<VillagerTrade> BUTCHER_LEVEL_2 = id("butcher/level_2");
+	TagKey<VillagerTrade> BUTCHER_LEVEL_2_2 = id("butcher/level_2_2");
+	TagKey<VillagerTrade> BUTCHER_LEVEL_3 = id("butcher/level_3");
+	TagKey<VillagerTrade> BUTCHER_LEVEL_4 = id("butcher/level_4");
+	TagKey<VillagerTrade> BUTCHER_LEVEL_5 = id("butcher/level_5");
+
+	TagKey<VillagerTrade> CARTOGRAPHER_LEVEL_1 = id("cartographer/level_1");
+	TagKey<VillagerTrade> CARTOGRAPHER_LEVEL_2 = id("cartographer/level_2");
+	TagKey<VillagerTrade> CARTOGRAPHER_LEVEL_3 = id("cartographer/level_3");
+	TagKey<VillagerTrade> CARTOGRAPHER_LEVEL_4 = id("cartographer/level_4");
+	TagKey<VillagerTrade> CARTOGRAPHER_LEVEL_5 = id("cartographer/level_5");
+
+	TagKey<VillagerTrade> CLERIC_LEVEL_1 = id("cleric/level_1");
+	TagKey<VillagerTrade> CLERIC_LEVEL_2 = id("cleric/level_2");
+	TagKey<VillagerTrade> CLERIC_LEVEL_3 = id("cleric/level_3");
+	TagKey<VillagerTrade> CLERIC_LEVEL_4 = id("cleric/level_4");
+	TagKey<VillagerTrade> CLERIC_LEVEL_5 = id("cleric/level_5");
+
+	TagKey<VillagerTrade> FARMER_LEVEL_1 = id("farmer/level_1");
+	TagKey<VillagerTrade> FARMER_LEVEL_2 = id("farmer/level_2");
+	TagKey<VillagerTrade> FARMER_LEVEL_3 = id("farmer/level_3");
+	TagKey<VillagerTrade> FARMER_LEVEL_4 = id("farmer/level_4");
+	TagKey<VillagerTrade> FARMER_LEVEL_5 = id("farmer/level_5");
+
+	TagKey<VillagerTrade> FISHERMAN_LEVEL_1 = id("fisherman/level_1");
+	TagKey<VillagerTrade> FISHERMAN_LEVEL_2 = id("fisherman/level_2");
+	TagKey<VillagerTrade> FISHERMAN_LEVEL_3 = id("fisherman/level_3");
+	TagKey<VillagerTrade> FISHERMAN_LEVEL_4 = id("fisherman/level_4");
+	TagKey<VillagerTrade> FISHERMAN_LEVEL_5 = id("fisherman/level_5");
+
+	TagKey<VillagerTrade> FLETCHER_LEVEL_1 = id("fletcher/level_1");
+	TagKey<VillagerTrade> FLETCHER_LEVEL_2 = id("fletcher/level_2");
+	TagKey<VillagerTrade> FLETCHER_LEVEL_3 = id("fletcher/level_3");
+	TagKey<VillagerTrade> FLETCHER_LEVEL_4 = id("fletcher/level_4");
+	TagKey<VillagerTrade> FLETCHER_LEVEL_5 = id("fletcher/level_5");
+
+	TagKey<VillagerTrade> LEATHERWORKER_LEVEL_1 = id("leatherworker/level_1");
+	TagKey<VillagerTrade> LEATHERWORKER_LEVEL_2 = id("leatherworker/level_2");
+	TagKey<VillagerTrade> LEATHERWORKER_LEVEL_3 = id("leatherworker/level_3");
+	TagKey<VillagerTrade> LEATHERWORKER_LEVEL_4 = id("leatherworker/level_4");
+	TagKey<VillagerTrade> LEATHERWORKER_LEVEL_5 = id("leatherworker/level_5");
+
+	TagKey<VillagerTrade> LIBRARIAN_LEVEL_1 = id("librarian/level_1");
+	TagKey<VillagerTrade> LIBRARIAN_LEVEL_2 = id("librarian/level_2");
+	TagKey<VillagerTrade> LIBRARIAN_LEVEL_3 = id("librarian/level_3");
+	TagKey<VillagerTrade> LIBRARIAN_LEVEL_4 = id("librarian/level_4");
+	TagKey<VillagerTrade> LIBRARIAN_LEVEL_5 = id("librarian/level_5");
+
+	TagKey<VillagerTrade> MASON_LEVEL_1 = id("mason/level_1");
+	TagKey<VillagerTrade> MASON_LEVEL_2 = id("mason/level_2");
+	TagKey<VillagerTrade> MASON_LEVEL_3 = id("mason/level_3");
+	TagKey<VillagerTrade> MASON_LEVEL_4 = id("mason/level_4");
+	TagKey<VillagerTrade> MASON_LEVEL_5 = id("mason/level_5");
+
+	TagKey<VillagerTrade> SHEPHERD_LEVEL_1 = id("shepherd/level_1");
+	TagKey<VillagerTrade> SHEPHERD_LEVEL_2 = id("shepherd/level_2");
+	TagKey<VillagerTrade> SHEPHERD_LEVEL_3 = id("shepherd/level_3");
+	TagKey<VillagerTrade> SHEPHERD_LEVEL_4 = id("shepherd/level_4");
+	TagKey<VillagerTrade> SHEPHERD_LEVEL_5 = id("shepherd/level_5");
+
+	TagKey<VillagerTrade> TOOLSMITH_LEVEL_1 = id("toolsmith/level_1");
+	TagKey<VillagerTrade> TOOLSMITH_LEVEL_2 = id("toolsmith/level_2");
+	TagKey<VillagerTrade> TOOLSMITH_LEVEL_3 = id("toolsmith/level_3");
+	TagKey<VillagerTrade> TOOLSMITH_LEVEL_4 = id("toolsmith/level_4");
+	TagKey<VillagerTrade> TOOLSMITH_LEVEL_5 = id("toolsmith/level_5");
+
+	TagKey<VillagerTrade> WEAPONSMITH_LEVEL_1 = id("weaponsmith/level_1");
+	TagKey<VillagerTrade> WEAPONSMITH_LEVEL_2 = id("weaponsmith/level_2");
+	TagKey<VillagerTrade> WEAPONSMITH_LEVEL_3 = id("weaponsmith/level_3");
+	TagKey<VillagerTrade> WEAPONSMITH_LEVEL_4 = id("weaponsmith/level_4");
+	TagKey<VillagerTrade> WEAPONSMITH_LEVEL_5 = id("weaponsmith/level_5");
 
 	static TagKey<VillagerTrade> id(String id) {
 		return TagKey.create(Registries.VILLAGER_TRADE, Stellarity.id(id));